refactor: Messages utils tests
